JSCBC-419: Updated to libcouchbase 2.8.0.
Reviewed-on: http://review.couchbase.org/83089
Tested-by: Brett Lawson <brett19@gmail.com>

+++ deps/lcb/tests/iotests/mock-environment.cc
@@ -112,7 +112,7 @@ std::vector<int> MockEnvironment::getMcPorts(std::string bucket)
 
     std::vector<int> ret;
 
-    for (int ii = 0; ii < payload.size(); ii++) {
+    for (int ii = 0; ii < (int)payload.size(); ii++) {
         ret.push_back(payload[ii].asInt());
     }
     return ret;
